refactor: restructure Equation Arena configuration and bonus logic
- Removed outdated configuration files and consolidated bonus logic into a new module.
- Moved animation styles to a dedicated directory for better organization.
