+++ src/lib/minigames/math/EquationArena/styles/animations.css
